[UIMA-6393]: Circular imports break resource manager cache
- Move import resolver test data to different test folder
- Better/added missing license header in the test files
