+++ uimaj-core/src/test/java/org/apache/uima/resource/metadata/impl/TypeSystemDescription_implTest.java
@@ -18,23 +18,36 @@
  */
 package org.apache.uima.resource.metadata.impl;
 
+import static java.util.Arrays.asList;
+import static java.util.stream.Collectors.toList;
+import static org.apache.uima.UIMAFramework.newDefaultResourceManager;
 import static org.apache.uima.test.junit_extension.JUnitExtension.getFile;
+import static org.apache.uima.util.CasCreationUtils.mergeTypeSystems;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.Assertions.tuple;
 
 import java.io.File;
+import java.net.URI;
+import java.nio.file.Paths;
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Stream;
 
 import org.apache.uima.UIMAFramework;
 import org.apache.uima.resource.ResourceInitializationException;
+import org.apache.uima.resource.ResourceManager;
 import org.apache.uima.resource.metadata.FeatureDescription;
 import org.apache.uima.resource.metadata.Import;
 import org.apache.uima.resource.metadata.TypeDescription;
 import org.apache.uima.resource.metadata.TypeSystemDescription;
+import org.apache.uima.test.junit_extension.JUnitExtension;
 import org.apache.uima.test.junit_extension.PrintExceptionsWhenRunFromCommandLineRule;
 import org.apache.uima.util.CasCreationUtils;
+import org.apache.uima.util.InvalidXMLException;
 import org.apache.uima.util.XMLInputSource;
 import org.apache.uima.util.XMLParser;
+import org.apache.uima.util.XMLizable;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Rule;
@@ -97,6 +110,80 @@ public class TypeSystemDescription_implTest {
                     tuple("testMultiRefAllowedFeature",
                             "A test feature that allows multiple references.", "uima.cas.FSArray",
                             null, true));
+  }
+
+  @Test
+  public void testResolveImports() throws Exception {
+    File descriptor = getFile("TypeSystemDescriptionImplTest/TestTypeSystem.xml");
+    TypeSystemDescription ts = xmlParser.parseTypeSystemDescription(new XMLInputSource(descriptor));
+
+    assertThat(ts.getTypes()).as("Type count before resolving the descriptor").hasSize(6);
+
+    assertThatThrownBy(() -> ts.resolveImports()).isInstanceOf(InvalidXMLException.class);
+    assertThat(ts.getTypes()).as(
+            "Type count after resolving failed should be same as before / no side effect on exception")
+            .hasSize(6);
+
+    // set data path correctly and it should work
+    ResourceManager resMgr = newDefaultResourceManager();
+    resMgr.setDataPath(
+            JUnitExtension.getFile("TypeSystemDescriptionImplTest/dataPathDir").getAbsolutePath());
+    ts.resolveImports(resMgr);
+
+    assertThat(ts.getTypes()) //
+            .as("Types after resolving the descriptor.") //
+            .extracting( //
+                    t -> Paths.get(URI.create(t.getSourceUrlString())).getFileName().toString(),
+                    TypeDescription::getName) //
+            .containsExactlyInAnyOrder( //
+                    tuple("TestTypeSystem.xml", "NamedEntity"), //
+                    tuple("TestTypeSystem.xml", "Person"), //
+                    tuple("TestTypeSystem.xml", "Place"), //
+                    tuple("TestTypeSystem.xml", "DocumentStructure"), //
+                    tuple("TestTypeSystem.xml", "Paragraph"), //
+                    tuple("TestTypeSystem.xml", "Sentence"), //
+                    tuple("TypeSystemImportedByName.xml", "TestType1"), //
+                    tuple("TypeSystemImportedByName.xml", "NamedEntity"), //
+                    tuple("TypeSystemImportedByName.xml", "TestType2"), //
+                    tuple("TypeSystemImportedByLocation.xml", "TestType1"), //
+                    tuple("TypeSystemImportedByLocation.xml", "TestType3"), //
+                    tuple("TypeSystemImportedFromDataPath.xml", "TestType4"), //
+                    tuple("TypeSystemImportedFromDataPath.xml", "TestType3"));
+    List<String> uniqueTypeNames = Stream.of(ts.getTypes()).map(TypeDescription::getName).distinct()
+            .sorted().collect(toList());
+    assertThat(uniqueTypeNames) //
+            .as("Unique type names after resolving the descriptor") //
+            .containsExactly("DocumentStructure", "NamedEntity", "Paragraph", "Person", "Place",
+                    "Sentence", "TestType1", "TestType2", "TestType3", "TestType4");
+    TypeSystemDescription mergedTsd = mergeTypeSystems(asList(ts));
+    assertThat(Stream.of(mergedTsd.getTypes()).map(TypeDescription::getName).sorted()
+            .collect(toList())).as("Type count after merging all the types ").hasSize(10);
+
+    String typeSystemImportedByLocation = new File(
+            "target/test-classes/TypeSystemDescriptionImplTest/TypeSystemImportedByLocation.xml")
+                    .toURI().toURL().toString();
+    String typeSystemImportedFromDataPath = new File(
+            "target/test-classes/TypeSystemDescriptionImplTest/dataPathDir/TypeSystemImportedFromDataPath.xml")
+                    .toURI().toURL().toString();
+    String typeSystemImportedByName = new File(
+            "target/test-classes/org/apache/uima/resource/metadata/impl/TypeSystemImportedByName.xml")
+                    .toURI().toURL().toString();
+
+    Map<String, XMLizable> cache = resMgr.getImportCache();
+    assertThat(cache).containsOnlyKeys(typeSystemImportedByLocation, typeSystemImportedByName,
+            typeSystemImportedFromDataPath);
+
+    TypeSystemDescription typeSystemImportedByLocationTSD = (TypeSystemDescription) cache
+            .get(typeSystemImportedByLocation);
+    assertThat(typeSystemImportedByLocationTSD.getTypes()).hasSize(2);
+
+    TypeSystemDescription typeSystemImportedFromDataPathTSD = (TypeSystemDescription) cache
+            .get(typeSystemImportedFromDataPath);
+    assertThat(typeSystemImportedFromDataPathTSD.getTypes()).hasSize(2);
+
+    TypeSystemDescription typeSystemImportedByNameTSD = (TypeSystemDescription) cache
+            .get(typeSystemImportedByName);
+    assertThat(typeSystemImportedByNameTSD.getTypes()).hasSize(3);
   }
 
   @Test
