--- uimaj-core/src/test/resources/TypeSystemDescriptionImplTest/Loop-with-3-nodes-1.xml
@@ -1,39 +1,0 @@
-<typeSystemDescription xmlns="http://uima.apache.org/resourceSpecifier">
-	
-        <!--
-	 ***************************************************************
-	 * Licensed to the Apache Software Foundation (ASF) under one
-	 * or more contributor license agreements.  See the NOTICE file
-	 * distributed with this work for additional information
-	 * regarding copyright ownership.  The ASF licenses this file
-	 * to you under the Apache License, Version 2.0 (the
-	 * "License"); you may not use this file except in compliance
-	 * with the License.  You may obtain a copy of the License at
-         *
-	 *   http://www.apache.org/licenses/LICENSE-2.0
-	 * 
-	 * Unless required by applicable law or agreed to in writing,
-	 * software distributed under the License is distributed on an
-	 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-	 * KIND, either express or implied.  See the License for the
-	 * specific language governing permissions and limitations
-	 * under the License.
-	 ***************************************************************
-   -->
-
-
-<name>Circular3-1</name>
-<description>For testing circular imports.</description>
-<version>0.1</version>
-<vendor>The Apache Software Foundation</vendor>
-<imports>
-<import location="Loop-with-3-nodes-2.xml"/>
-</imports>
-<types>
-<typeDescription>
-<name>Foo</name>
-<description/>
-<supertypeName>uima.tcas.Annotation</supertypeName>
-</typeDescription>
-</types>
-</typeSystemDescription>
