--- uimaj-core/src/test/resources/TypeSystemDescriptionImplTest/complexImportScenario1/tsd1.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<typeSystemDescription xmlns="http://uima.apache.org/resourceSpecifier">
-    <imports>
-        <import location="tsd6.xml"/>
-    </imports>
-    <types>
-        <typeDescription>
-            <name>Type1_0</name>
-            <description/>
-            <supertypeName>uima.cas.TOP</supertypeName>
-            <features>
-                <featureDescription>
-                    <name>Feature_1_0_0</name>
-                    <description/>
-                    <rangeTypeName>uima.cas.Integer</rangeTypeName>
-                </featureDescription>
-            </features>
-        </typeDescription>
-    </types>
-</typeSystemDescription>
