--- uimaj-core/src/test/resources/TypeSystemDescriptionImplTest/complexImportScenario1/tsd2.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<typeSystemDescription xmlns="http://uima.apache.org/resourceSpecifier">
-    <imports>
-        <import location="tsd3.xml"/>
-        <import location="tsd6.xml"/>
-        <import location="tsd5.xml"/>
-        <import location="tsd1.xml"/>
-    </imports>
-    <types>
-        <typeDescription>
-            <name>Type2_0</name>
-            <description/>
-            <supertypeName>uima.cas.TOP</supertypeName>
-            <features>
-                <featureDescription>
-                    <name>Feature_2_0_0</name>
-                    <description/>
-                    <rangeTypeName>uima.cas.Integer</rangeTypeName>
-                </featureDescription>
-            </features>
-        </typeDescription>
-    </types>
-</typeSystemDescription>
