--- uimaj-core/src/test/resources/TypeSystemDescriptionImplTest/complexImportScenario1/tsd5.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<typeSystemDescription xmlns="http://uima.apache.org/resourceSpecifier">
-    <imports>
-        <import location="tsd2.xml"/>
-        <import location="tsd6.xml"/>
-    </imports>
-    <types>
-        <typeDescription>
-            <name>Type5_0</name>
-            <description/>
-            <supertypeName>uima.cas.TOP</supertypeName>
-            <features>
-                <featureDescription>
-                    <name>Feature_5_0_0</name>
-                    <description/>
-                    <rangeTypeName>uima.cas.Integer</rangeTypeName>
-                </featureDescription>
-            </features>
-        </typeDescription>
-    </types>
-</typeSystemDescription>
