--- uimaj-core/src/test/resources/TypeSystemDescriptionImplTest/complexImportScenario1/tsd6.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<typeSystemDescription xmlns="http://uima.apache.org/resourceSpecifier">
-    <types>
-        <typeDescription>
-            <name>Type6_0</name>
-            <description/>
-            <supertypeName>uima.cas.TOP</supertypeName>
-            <features>
-                <featureDescription>
-                    <name>Feature_6_0_0</name>
-                    <description/>
-                    <rangeTypeName>uima.cas.Integer</rangeTypeName>
-                </featureDescription>
-            </features>
-        </typeDescription>
-    </types>
-</typeSystemDescription>
